+++ src/main/groovy/tradingmaster/model/ITrade.groovy
@@ -3,8 +3,11 @@ package tradingmaster.model
 interface  ITrade {
 
     String getExtId()
-    Double getQuantity()
+
+    BigDecimal getQuantity()
+
     BigDecimal getPrice()
+
     Date getDate()
 
 }
